chore: clarify and consolidate management repo classes
AM-6696
