+++ gravitee-am-repository/gravitee-am-repository-mongodb/src/main/java/io/gravitee/am/repository/mongodb/management/internal/model/GroupMongo.java
@@ -26,7 +26,6 @@ import java.util.Objects;
  * @author GraviteeSource Team
  */
 public class GroupMongo extends Auditable {
-    // TODO [DP] class to remove
     @BsonId
     private String id;
